added in bq25883 driver definitions and added in the ability to read all registers from charger chip

+++ PainDrain/PainDrain.cydsn/tens.h
@@ -17,7 +17,7 @@ void set_tens_task(void);
 void set_tens_freq(double period);
 void set_tens_dur(int duration, int tens_dur_ms);
 void set_tens_amp(int amplitude);
-void set_tens_signal(int amplitude, double duration, double period, int channel, int phase);
+void set_tens_signal(int amplitude, double duration, int period, int channel, int phase);
 
 #endif /* TENS_H */
 /* [] END OF FILE */
